refactor: `queryStatistics` as standalone fun
