+++ app/main.js
@@ -1,4 +1,5 @@
 import { changeMenuIcon, showSidePanel } from "./ui";
+import { queryStatistics } from "./query-zoning";
 import { zoningNumberChart, zoningAreaChart, updateChart, clearCharts } from "./create-chart";
 
 
@@ -334,7 +335,7 @@ const debouncedRunQuery = debounce(function () {
   calculateAreaByZoning();
 
   return eachAlways([
-    queryStatistics(),
+    queryStatistics(featureToQuery, sketchGeometry, bufferSize),
     updateMapLayer()
   ]);
 });
@@ -585,76 +586,6 @@ function updateMapLayer () {
   return featureToQuery.queryObjectIds(query).then(changeFeatureCount);
   // return webLayerView.queryObjectIds(query).then(highlightGeometries);
 }
-
-const statDefinitions = [{
-  onStatisticField: "CASE WHEN ZONE_MAS = 'R(A)' THEN 1 ELSE 0 END",
-  outStatisticFieldName: "zone_RA",
-  statisticType: "sum"
-},
-{
-  onStatisticField: "CASE WHEN ZONE_MAS = 'R(B)' THEN 1 ELSE 0 END",
-  outStatisticFieldName: "zone_RB",
-  statisticType: "sum"
-},
-{
-  onStatisticField: "CASE WHEN ZONE_MAS = 'R(C)' THEN 1 ELSE 0 END",
-  outStatisticFieldName: "zone_RC",
-  statisticType: "sum"
-},
-{
-  onStatisticField: "CASE WHEN ZONE_MAS = 'G/IC' THEN 1 ELSE 0 END",
-  outStatisticFieldName: "zone_GIC",
-  statisticType: "sum"
-},
-{
-  onStatisticField: "CASE WHEN ZONE_MAS = 'O' THEN 1 ELSE 0 END",
-  outStatisticFieldName: "zone_O",
-  statisticType: "sum"
-},
-{
-  onStatisticField: "CASE WHEN ZONE_MAS = 'C' THEN 1 ELSE 0 END",
-  outStatisticFieldName: "zone_C",
-  statisticType: "sum"
-},
-{
-  onStatisticField: "CASE WHEN ZONE_MAS = 'MRDJ' THEN 1 ELSE 0 END",
-  outStatisticFieldName: "zone_MRDJ",
-  statisticType: "sum"
-},
-{
-  onStatisticField: "CASE WHEN ZONE_MAS NOT IN ('R(A)', 'R(B)', 'R(C)', 'G/IC', 'O', 'C', 'MRDJ') THEN 1 ELSE 0 END",
-  outStatisticFieldName: "zone_OTHERS",
-  statisticType: "sum"
-}
-];
-
-function queryStatistics () {
-  // https://developers.arcgis.com/javascript/latest/api-reference/esri-tasks-support-Query.html
-  const query = featureToQuery.createQuery();
-
-  query.geometry = sketchGeometry;
-  query.distance = bufferSize;
-  query.outStatistics = statDefinitions;
-
-  // with outStatistics returned result is a "table" with geometry of null
-  return featureToQuery.queryFeatures(query).then(function (result) {
-    // console.log(result);
-
-    const allStats = result.features[0].attributes;
-
-    updateChart(zoningNumberChart, [
-      allStats.zone_RA,
-      allStats.zone_RB,
-      allStats.zone_RC,
-      allStats.zone_GIC,
-      allStats.zone_O,
-      allStats.zone_C,
-      allStats.zone_MRDJ,
-      allStats.zone_OTHERS
-    ]);
-  }, console.error);
-}
-
 
 /* sidebar */
